Fixed https://github.com/samboy/ObHack/issues/3
This was tricky: The method that initially draws the box is a private
method that must be called *before* we load CONFIG.cfg.  Fortunately,
I did not have to write a new method that just redraws the values in
the UI box after we load CONFIG.cfg; instead, I am able to call the
method Oblige uses to lock the values while building a level to unlock
the values; this has the side effect of correctly locking or unlocking
the maxsize value.

Next: I need to make sure the Enclosures value is correctly locked or
unlocked depending on whether we make a Heretic/Blasphemer megawad, and
I think it makes sense to never lock “key type”, since, in Heretic, that
option can only have one value.