The issue with monsters going through the ceiling was simple
It was a clerical error: cage_medium_liq_LAVA did not have min_height
However, it has changed some of the maps (the blue key room in MAP28
now has three Revenants as guards instead of the old Baron and mix
of zombies, shooters, and imps), so time to update the tests.

Also, I start updating the tests for the modified levels.

Note that while MAP29 and MAP30 appear to be different, they play the same;
it is only the Blue Key room in MAP28 which has changed.

The 1FreeDoom1.test.sh script makes sure the maps are completely identical

The 1FreeDoom1-sp.test.sh script makes sure the maps are the same in
single player mode.  Note that, in this test, only MAP28 was altered,
which indicates the only change in MAP29 and MAP30 is the sorting of
things (perhaps I should update 1FreeDoom1.test.sh to sort things the
way the “sp” test does; note that the sorting requires using non-POSIX
GAWK-specific features).

The 1FreeDoom1-game.test.sh, 1FreeDoom1-mode.test.sh and
1FreeDoom1-monsters.test.sh scripts use the same SHA-256 sums.

See also: https://github.com/samboy/ObHack/issues/10