Modify DOOM2.WAD to use mostly .MUS files
This fixes https://github.com/samboy/ObHack/issues/15

When I played a map, then hit the exit and moved on to the next map,
sometimes the MIDI (song) for the new map did not start.

In particular, it happened when:

*   Going from MAP21 (D_COUNT2) to MAP22 (D_DDTBL3)
*   Going from MAP24 (D_THEDA3) to MAP25 (D_ADRIAN)
*   Going from MAP27 (D_ROMER2) to MAP28 (D_TENSE)

I could reliably reproduce the issue with the MAP25 music reproducing
using the default instant-exit levels and starting around level 19.

Converting the newer Freedoom .mid files in to .mus files appears to
solve the problem; I can no longer reproduce the issue after making
the conversion. In addition, the converted .mus files seamlessly loop
(there’s a pause with the .mid form of the same songs) and sound like
they play without issue after being converted.

The MAP09 (D_IN_CIT) and MAP21 (D_COUNT2) files need to remain .mid files
because the files get corrupted when converted in to .mus files. There
appears to be no issue starting these music files, so I have made an
updated DOOM2.WAD and close this issue.

I have also added the tool to convert .mid files in to .mus files to
the Git tree.

This released ObHack-FreeDoom-3.3.9

Hopefully, the music issues are solved once and for all.