ObHack 3.6 update: Update the tests
Since I have fixed the “move deathmatch goodies” bug where a change
to a given level would move the deathmatch goodies in all subsequent
levels, this altered where the goodies were in a lot of maps.

Hence, the tests are updated.

The good news is that a future change which affects only, say, MAP02
will only affect that one map when updating the tests.